Fix x64 runtime issues

+++ src/xrCore/rt_compressor.h
@@ -3,8 +3,8 @@
 #pragma once
 
 extern XRCORE_API void	rtc_initialize	();
-extern XRCORE_API u32		rtc_compress	(void *dst, u32 dst_len, const void* src, u32 src_len);
-extern XRCORE_API u32		rtc_decompress	(void *dst, u32 dst_len, const void* src, u32 src_len);
+extern XRCORE_API size_t rtc_compress	(void *dst, u32 dst_len, const void* src, u32 src_len);
+extern XRCORE_API size_t rtc_decompress	(void *dst, u32 dst_len, const void* src, u32 src_len);
 extern XRCORE_API u32		rtc_csize		(u32 in);
 
 extern XRCORE_API void      rtc9_initialize	();
